update to use websockets on template service

+++ android-app/app/src/main/java/com/example/androidapp/components/feed/FeedViewModel.kt
@@ -3,21 +3,24 @@ package com.example.androidapp.components.feed
 import android.content.Context
 import android.util.Log
 import androidx.lifecycle.*
+import com.example.androidapp.TEMPLATE_ENDPOINT
 import com.example.androidapp.data.Feed
 import com.example.androidapp.data.FeedDatabase
 import com.example.androidapp.data.FeedRepo
-import com.example.androidapp.services.RemoteDataSource
-import com.example.androidapp.signals.SignalProvider
+import com.example.androidapp.models.factories.FeedResponseFactory
+import com.example.androidapp.services.MessageListener
+import com.example.androidapp.services.WebSocketManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.launch
+import org.json.JSONArray
 import javax.inject.Inject
 
 @HiltViewModel
 class FeedViewModel @Inject constructor(
     @ApplicationContext context: Context,
-    private val remoteDataSource: RemoteDataSource
-): ViewModel() {
+    private val feedResponseFactory: FeedResponseFactory
+): ViewModel(), MessageListener {
 
     private val _feed = MutableLiveData<Feed>()
     val feed: LiveData<Feed> = _feed
@@ -28,16 +31,35 @@ class FeedViewModel @Inject constructor(
         val feedDao = FeedDatabase.getDatabase(context).feedDao()
         repo = FeedRepo(feedDao)
         _feed.value = repo.readAllData.value
-        getResponse()
+        val ws = WebSocketManager
+        ws.init(TEMPLATE_ENDPOINT, this)
+        ws.connect()
     }
 
-    private fun getResponse() {
-        viewModelScope.launch {
-            remoteDataSource.getFeed() { it ->
-                val feedResponse = Feed(0, it)
+    override fun onConnectSuccess() {
+        Log.d("INFO", "Connected to template service web socket")
+    }
+
+    override fun onConnectFailed() {
+        Log.e("ERROR", "Failed to connect to template service web socket")
+    }
+
+    override fun onClose() {
+        Log.d("INFO", "Connection to template service has closed")
+    }
+
+    override fun onMessage(text: String?) {
+        if (text != null) {
+            val json = JSONArray(text)
+            if (json.length() > 0) {
+                val feed = json.getJSONObject(0).getJSONObject("data").getJSONArray("elements")
+                val feedResponse = Feed(0, feedResponseFactory.create(feed))
                 _feed.postValue(feedResponse)
-                repo.deleteFeed()
-                repo.addFeed(feedResponse)
+
+                viewModelScope.launch {
+                    repo.deleteFeed()
+                    repo.addFeed(feedResponse)
+                }
             }
         }
     }
